ALDE-2859 filtering consumers

+++ app/Repositories/UserRepository.php
@@ -7,12 +7,16 @@ use App\Http\Resources\UserResource;
 use App\User;
 use App\QueryBuilders\UserSearch;
 use App\UserInfo;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Pipeline\Pipeline;
 use bigfood\grid\RepositoryInterface;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class UserRepository implements RepositoryInterface
 {
+    use AuthorizesRequests;
+
     /** @var User */
     protected $model;
 
@@ -46,6 +50,8 @@ class UserRepository implements RepositoryInterface
      */
     public function add(array $data)
     {
+        $this->authorize('create', User::class);
+
         return new UserResource($this->model->create($data));
     }
 
@@ -58,6 +64,9 @@ class UserRepository implements RepositoryInterface
     {
         /** @var User $model */
         $model = $this->model->findOrFail($id);
+
+        $this->authorize('update', $model);
+
         $model->update($data);
 
         $model->userInfo->update($data['user_info']);
@@ -71,7 +80,23 @@ class UserRepository implements RepositoryInterface
      */
     public function delete($id)
     {
+        $this->authorize('delete', $this->model);
+
         return $this->model->destroy($id);
+    }
+
+    /**
+     * @return array
+     */
+    public function getAllowActions()
+    {
+        return [
+            'all'    => true,
+            'create' => true,
+            'view'   => Gate::allows('view', $this->model),
+            'edit'   => true,
+            'delete' => true,
+        ];
     }
 
     /**
@@ -80,6 +105,10 @@ class UserRepository implements RepositoryInterface
      */
     public function get($id)
     {
-        return new UserResource($this->model->with('userInfo')->findOrFail($id));
+        $model = $this->model->with('userInfo')->findOrFail($id);
+
+        $this->authorize('view', $model);
+
+        return new UserResource($model);
     }
 }
